Refactor SIMD modules
- Removes the CpuFeatureLevel struct and params, which added complexity
  for no real world benefit
- The new structure results in smaller, more organized files
- The new structure will also be more testable, although new tests are
  not yet added
