+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "av-scenechange"
-version = "0.14.1"
+version = "0.15.0"
 authors = ["Josh Holmer <jholmer.in@gmail.com>"]
 edition = "2021"
 rust-version = "1.74.1"
@@ -14,8 +14,10 @@ aligned = "0.4.2"
 anyhow = "1.0.56"
 arg_enum_proc_macro = "0.3.4"
 arrayvec = "0.7.6"
+cfg-if = "1.0.0"
 clap = { version = "4.0.22", optional = true, features = ["derive"] }
 console = { version = "0.15", optional = true }
+cpufeatures = "0.2.17"
 fern = { version = "0.7", optional = true }
 ffmpeg-the-third = { version = "3.0.0", optional = true, default-features = false, features = [
     "codec",
@@ -42,7 +44,6 @@ vapoursynth = { version = "0.4.0", features = [
 ], optional = true }
 y4m = "0.8.0"
 
-
 [build-dependencies]
 cc = { version = "1.2.23", optional = true, features = ["parallel"] }
 nasm-rs = { version = "0.3", optional = true, features = ["parallel"] }
@@ -55,7 +56,6 @@ devel = ["console", "fern"]
 tracing = ["tracing-subscriber", "tracing-chrome", "dep:tracing"]
 ffmpeg = ["ffmpeg-the-third"]
 asm = ["nasm-rs", "cc", "libc"]
-libc = ["dep:libc"]
 
 [[bin]]
 name = "av-scenechange"
